fix xds_bootstrap_test
